Overview map setup - reworking IP2024-21
Instead of a static picture of the floor, now it is a camera that watches the floor.
Added pinching and dragging functionality.
Switching between AR Mode and Map Overview works smoothly.
Exit Nav works smoothly.
After entering Map Overview, information on the top of the screen of the current target location is shown.